Fixed Metric Conversion Error
